+++ protocol/src/entities/database/member.rs
@@ -14,19 +14,19 @@ pub struct GroupToPeer {
 }
 
 pub trait GroupMember {
-    fn list_members(&self, id: u8) -> Result<Vec<Peer>, Box<dyn Error>>;
+    fn list_members(&self, id: u16) -> Result<Vec<Peer>, Box<dyn Error>>;
     fn insert_member(
         &self,
-        group_idx: u8,
+        group_idx: u16,
         member: Peer,
         known: bool,
     ) -> Result<Peer, Box<dyn Error>>;
     fn list_groups_with_member(&self, peer_id: u16) -> Result<Vec<DBGroup>, Box<dyn Error>>;
-    fn remove_member(&self, peer_id: u32, group_id: u32) -> Result<(), Box<dyn Error>>;
+    fn remove_member(&self, peer_id: u16, group_id: u16) -> Result<(), Box<dyn Error>>;
 }
 
 impl GroupMember for Connection {
-    fn list_members(&self, group_id: u8) -> Result<Vec<Peer>, Box<dyn Error>> {
+    fn list_members(&self, group_id: u16) -> Result<Vec<Peer>, Box<dyn Error>> {
         let mut stmt = self.prepare("SELECT * FROM group_to_peer WHERE group_id = ?1")?;
         let rows = stmt.query_map([group_id], |r| {
             Ok(GroupToPeer {
@@ -53,7 +53,7 @@ impl GroupMember for Connection {
         let mut result = vec![];
         for r in rows {
             let r = r?;
-            let group = self.get_group_by_idx(r as u8)?;
+            let group = self.get_group_by_idx(r as u16)?;
             result.push(group);
         }
         Ok(result)
@@ -61,7 +61,7 @@ impl GroupMember for Connection {
 
     fn insert_member(
         &self,
-        group_idx: u8,
+        group_idx: u16,
         mut member: Peer,
         known: bool,
     ) -> Result<Peer, Box<dyn Error>> {
@@ -78,7 +78,7 @@ impl GroupMember for Connection {
         }
     }
 
-    fn remove_member(&self, peer_id: u32, group_id: u32) -> Result<(), Box<dyn Error>> {
+    fn remove_member(&self, peer_id: u16, group_id: u16) -> Result<(), Box<dyn Error>> {
         let mut stmt =
             self.prepare("DELETE FROM group_to_peer WHERE peer_id = ?1 AND group_id = ?2")?;
         stmt.execute(params![peer_id, group_id])?;
